svm: remove bogus MSRPM test
The MSR permission bitmap is two pages long.  Therefore, the first test
in test_msrpm_iopm_bitmap_addrs should actually pass.  The problem is
that KVM tries to merge the L1 MSR permission bitmap with its own at
VMRUN time, fails to map the address of the L1 MSR permission bitmap,
and fails.

Thus remove the bogus test, which is failing on QEMU.

Reported-by: Lara Lazier <laramglazier@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>