s390x: snippets: Add gitignore as well as linker script and start assembly
Snippets are small guests which can be run under a unit test that
implements a simple hypervisor. They can be written in C or
assembly. The C code needs a linker script and a start assembly file
that jumps to main to work properly. So let's add that as well as a
gitignore entry for the new files.

Signed-off-by: Janosch Frank <frankja@linux.ibm.com>
Reviewed-by: Thomas Huth <thuth@redhat.com>