Test compilation with Clang on aarch64, ppc64le and s390x in Travis-CI
Travis-CI recently changed their policy so that builds on the non-x86
build machines are possible without consuming any credits again. We can
use these systems to test compilation of the non-x86 code with Clang.
Unfortunately, the qemu-system-s390x of Ubuntu 20.04 seems to be buggy,
so that the s390x binaries cause that QEMU to crash. Thus we can only
run the TCG tests for ppc64le and aarch64.

Message-Id: <20210622135517.234801-5-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>