s390x: Fix out-of-tree builds
The support for "snippets" (nested guest binaries) that has been added
recently to the s390x folder broke the out-of-tree compilation. We
have to make sure that the snippet folder is created in the build
directory, too, and that linker script is taken from the source folder.

While we're at it, switch the gitlab-CI cross compiler job to use
out-of-tree builds, too, so that this does not happen so easily again.
We're still testing in-tree s390x builds with the native "s390x-kvm"
job on the s390x host, so we now test both, in-tree and out-of-tree
builds.

Fixes: 2f6fdb4ac446 ("s390x: snippets: Add snippet compilation")
Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <20210716105219.1201997-1-thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>