travis.yml: Disable the failing gicv2-mmio test
The test is constantly failing now. After trying the old commit
that re-introduced the .travis.yml file again in Travis-CI, it is
now also failing there, so this is likely not a regression in the
kvm-unit-tests, but a regression in the Travis-CI / Ubuntu build
environment. Thus let's simply disable the test for now, since
there is not much else we can do from the kvm-unit-test side here.

Message-Id: <20210722172354.3759847-1-thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>