arm64: Link with -z notext
Building the tests on arm64 fails when using LLD (the LLVM linker):

  ld.lld: error: can't create dynamic relocation R_AARCH64_ABS64 against
  local symbol in readonly segment; recompile object files with -fPIC or
  pass '-Wl,-z,notext' to allow text relocations in the output
  >>> defined in lib/libcflat.a(processor.o)
  >>> referenced by processor.c
  >>>               processor.o:(vector_names) in archive lib/libcflat.a

The reason is that LLD defaults to errors for text relocations. The GNU
LD defaults to let it go. In fact, the same error can be reproduced when
using GNU LD with the '-z text' arg (to error on text relocations):

  aarch64-linux-gnu-ld: read-only segment has dynamic relocations

Fix this link error by adding `-z notext` into the arm64 linker flags.

Signed-off-by: Ricardo Koller <ricarkol@google.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>