x86/msr.c generalize to any input msr
If an MSR description is provided as input by the user,
run the test against that MSR. This allows the user to
run tests on custom MSR's.

Otherwise run all default tests.

This is to validate custom MSR handling in user space
with an easy-to-use tool. This kvm-unit-test submodule
is a perfect fit. I'm extending it with a mode that
takes an MSR index and a value to test arbitrary MSR accesses.

Signed-off-by: Daniele Ahmed <ahmeddan@amazon.com>
Message-Id: <20210927153028.27680-3-ahmeddan@amazon.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>