x86: SEV-ES: add port string IO test case
Add a test case to verify that string IO works as expected under SEV-ES.
This test case is based on the `test_stringio()` test case in emulator.c.
However, emulator.c does not currently run under UEFI.

Only the first half of the test case, which processes a string from
beginning to end, was taken for now. The second test case did not work
and is thus left out of the amd_sev.c setup for now.

Also, the first test case was modified to do port IO at word granularity
rather than byte granularity. The reason is to ensure that using the
port IO size in a calculation within the kernel does not multiply or
divide by 1. In particular, this tweak is useful to demonstrate that a
recent KVM patch [1] does not behave correctly.

* This patch is based on the `uefi` branch.

[1] https://patchwork.kernel.org/project/kvm/patch/20211013165616.19846-2-pbonzini@redhat.com/

Signed-off-by: Marc Orr <marcorr@google.com>
Message-Id: <20211025052829.2062623-1-marcorr@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>