x86: Look up the PTEs rather than assuming them
Rather than assuming which PTEs the SMEP test runs on, look them up to
ensure they are correct.  If this test were to run on a different page
table (ie: run in an L2 test) the wrong PTEs would be set.  Switch to
looking up the PTEs to avoid this from happening.

Signed-off-by: Aaron Lewis <aaronlewis@google.com>
Message-Id: <20211110212001.3745914-11-aaronlewis@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>