arch-run: do not process ERRATA when running under TCG
All the errata checking looks at the current host kernel version. For
TCG runs this is entirely irrelevant as the host kernel has no impact
on the behaviour of the guest. In fact we should set ERRATA_FORCE to
ensure we run those tests as QEMU doesn't attempt to model
non-confirming architectures.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
[Removed unnecessary eval from the export ERRATA_FORCE line.]
Message-Id: <20211112114734.3058678-4-alex.bennee@linaro.org>
Signed-off-by: Andrew Jones <drjones@redhat.com>