arm/flat.lds: don't drop debug during link
It is useful to keep the debug in the .elf file so we can debug and it
doesn't get copied across to the final .flat file. Of course we still
need to ensure we apply the offset when we load the symbols based on
where QEMU decided to load the kernel.

  (gdb) symbol-file ./builds/arm64/arm/tlbflush-data.elf -o 0x40080000

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>
Message-Id: <20211202115352.951548-3-alex.bennee@linaro.org>
Signed-off-by: Andrew Jones <drjones@redhat.com>