x86: Overhaul definitions for DR6 and DR7 bits
Clean up the mess that is debugreg.h to follow the nomenclature used by
the SDM and the kernel (as best as possible).  Use the "new" defines in
various tests.  Opportunistically add a define for VMX's extra flag in
vmcs.GUEST_PENDING_DBG_EXCEPTIONS that is set if any DR0-3 trap matched
and was enabled.

No functional change intended.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Message-Id: <20220120002923.668708-6-seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>