s390x: firq: fix running in PV
If using the qemu CPU type, Protected Virtualization is not available,
and the test will fail to start when run in PV. If specifying the host
CPU type, the test will fail to start with TCG because the host CPU
type requires KVM. In both cases the test will show up as failed.

This patch adds a copy of the firq test definitions for KVM, using the
host CPU type, and specifying that KVM is to be used. The existing
firq tests are then fixed by specifying that TCG is to be used.

When running the tests normally, both variants will be run. If an
accelerator is specified explicitly when running the tests, only one
variant will run and the other will be skipped (and not fail).

Signed-off-by: Claudio Imbrenda <imbrenda@linux.ibm.com>
Fixes: 8b98745d ("s390x: firq: floating interrupt test")
Acked-by: David Hildenbrand <david@redhat.com>
Acked-by: Janosch Frank <frankja@linux.ibm.com>