x86/pmu: Set appropriate expectations for reference cycles
Per the SDM, reference cycles (as a general purpose event) are
currently implemented using the core crystal clock, TSC, or bus
clock. Therefore, it's unreasonable to expect reference cycles for the
measured loop to fall within some constant multiples of the number of
loop iterations, unless those bounds are set so wide as to be
pointless.

The bounds initially established when this test was written were
broadened in commit 4779578b24b3 ("make PMU test to pass on more cpu
types"), but even the new bounds are too narrow to accommodate a
2.6GHz Ice Lake, with the TSC frequency at 104 times the reference
cycle (for this implementation, the core crystal clock) frequency.

Restore the initial (tighter) bounds, calculate the ratio of TSC
frequency to reference cycle frequency, and then scale the bounds
accordingly.

Tested on several generations of Xeon E5 parts: Ice Lake, Cascade
Lake, Skylake, Broadwell, and Haswell.

Opportunistically fixed a spelling error and a commented-out printf
format string.

Fixes: 4779578b24b3 ("make PMU test to pass on more cpu types")
Reported-by: Jacob Xu <jacobhxu@gmail.com>
Signed-off-by: Jim Mattson <jmattson@google.com>
Message-Id: <20220213082714.636061-1-jmattson@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>